ID3v2: Move `Id3v2Tag` unit tests to `tag/tests.rs`
We reached nearly 1.5k lines of unit tests for `Id3v2Tag`, and with the `tag` module already being quite large itself, it was nearly 3k lines total.

Moving the unit tests to `{module}/tests.rs` is a common pattern in std.
